Improved RFI Detection + Med Sub
Uses Chauvenet's criterion to determine the nsigma above which data are flagged as outliers. It then uses the median absolute deviation (MAD) to flag the data by the following method which assumes the data are normally distributed:
abs(data-median(data)) / 1.4826*MAD
where the MAD = median(abs(data-median(data)))
It then flags any absolute values of this ratio if they are greater then the tolerance level found using Chauvenet's criterion.
The complex median is then taken of this data, and within the function called to write the new uv file, the data are flagged and the median subtracted.