Flagger and Subtractor
Script with the option to flag data using Chauvenet's criterion and then subtract either the mean or median over all files in directory. If the flagging option is not chosen, the user can choose to subtract the mean or median calculated using all the files in a directory from each uv file read in. Each file is then saved as a new uv file with one of the following additions to the file extension:
m: median subtracted (e.g. .uvm)
me: mean subtracted
Rm: flagged using median; median is also subtracted
Rme: flagged using mean; mean is also subtracted
User is prompted for the data directory they wish to process