Directory-based Statistics
Now takes an entire directory of UV files, builds a data array, and calculates the mean, median, and maximum values before and after flagging (using either the mean or median and Chauvenet's criterion). Also calculates the difference between the pre- and post-flagged data mean and median. It plots all of these.

Changed plotting colors and linestyles for some of the plots to account for colorblindness and make it easier to distinguish between data curves.